Node Monitor Refactoring  - Fetch status renamed to main.  - Fetch status library renamed to fetch status.  - Fetching of pool connection and fetch status are now classes.  - Created util file for helper functions.  - Fixed and added logging.  - Removed some imports and requirements.
