+++ fetch-validator-status/Dockerfile
@@ -20,4 +20,4 @@ RUN pip install --no-cache-dir -r requirements.txt
 ADD networks.json .
 ADD *.py ./
 
-ENTRYPOINT ["bash", "-c", "python fetch_status.py $@", "--"]
+ENTRYPOINT ["bash", "-c", "python main.py $@", "--"]
